+++ react/game/src/app/components/Logo/styles.scss
@@ -7,10 +7,10 @@
 @keyframes logo {
   from {
     opacity: 0;
-    transform: scale(0.1, 0.1) rotate3d(0, 0.9, 0.5, 360deg);
+    transform: scale(0.1) rotate3d(0, 0.9, 0.5, 360deg);
   }
   to {
     opacity: 1;
-    transform: scale(1, 1) rotate3d(0, 0, 0, 0);
+    transform: scale(1) rotate3d(0, 0, 0, 0);
   }
 }
