--- react/game/src/constants/redux.js
@@ -1,13 +1,0 @@
-export const GAME_ACTIONS = {
-  GAME_PLAYER_MOVED: 'GAME_PLAYER_MOVED',
-  GAME_HISTORY_CHANGED: 'GAME_HISTORY_CHANGED',
-  LOAD_GAME_INFO: 'LOAD_GAME_INFO'
-};
-
-export const LOGIN_ACTIONS = {
-  LOAD_APP: 'LOAD_APP',
-  APP_LOADED: 'APP_LOADED',
-  LOGIN_LOADING: 'LOGIN_LOADING',
-  LOGIN_ERROR: 'LOGIN_ERROR',
-  LOGIN_SUCCESS: 'LOGIN_SUCCESS'
-};
