+++ react/game/src/services/authService.js
@@ -2,5 +2,5 @@ import { API } from '@config/api';
 import { API_LOGIN_PATH } from '@constants/apiNames';
 
 export const service = {
-  post: values => API.post(API_LOGIN_PATH, { email: values.email, password: values.password })
+  post: ({ email, password }) => API.post(API_LOGIN_PATH, { email, password })
 };
