+++ react/game/src/utils/router.js
@@ -1,5 +1,2 @@
 export const checkRoute = (routeToCheck, pathname) =>
-  Object.keys(routeToCheck).some(key => {
-    if (routeToCheck[key].path === pathname) return true;
-    return false;
-  });
+  Object.keys(routeToCheck).some(key => routeToCheck[key].path === pathname);
