+++ react/game/src/validation/forms/index.js
@@ -1,14 +1,15 @@
 /* eslint-disable no-useless-escape */
 const emailRegex = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
 /* eslint-enable */
+const INPUT_NAME_INDEX = 2;
 
 export const required = (value, ...rest) => {
-  const inputName = rest[2];
+  const inputName = rest[INPUT_NAME_INDEX];
   return value ? undefined : `${inputName} is required`;
 };
 
 export const minLength = (value, ...rest) => {
-  const inputName = rest[2];
+  const inputName = rest[INPUT_NAME_INDEX];
   return value.length < 8 ? `${inputName} must be at least 8 characters` : undefined;
 };
